包结构调整,注释完善
