+++ src/main/java/cn/wxingzou/conditionflow/result/FailReason.java
@@ -1,0 +1,14 @@
+package cn.wxingzou.conditionflow.result;
+
+/**
+ * 失败原因
+ *
+ * @author wuxiaolin
+ * @version 1.0
+ * @date 2019/11/8 17:36
+ **/
+public interface FailReason<R> {
+
+    R getReason();
+
+}
