--- src/main/java/cn/wxingzou/flowengine/result/FailReason.java
@@ -1,12 +1,0 @@
-package cn.wxingzou.flowengine.result;
-
-/**
- * @author wuxiaolin
- * @version 1.0
- * @date 2019/11/8 17:36
- **/
-public interface FailReason<R> {
-
-    R getReason();
-
-}
