--- src/test/java/cn/wxingzou/flowengine/GroupbookingApplicationTests.java
@@ -1,13 +1,0 @@
-package cn.wxingzou.flowengine;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GroupbookingApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
